+++ eeg_flow/config/__init__.py
@@ -1,3 +1,4 @@
 """Configuration module."""
 
-from .config import create_config, load_config, load_triggers  # noqa: F401
+from .config import create_config, load_config  # noqa: F401
+from .triggers import load_triggers  # noqa: F401
