+++ eeg_flow/config/config.py
@@ -14,41 +14,6 @@ if TYPE_CHECKING:
 _TRIGGER_FNAME = files("eeg_flow.config") / "triggers.ini"
 
 
-def load_triggers(fname: Union[str, Path] = _TRIGGER_FNAME) -> dict[str, int]:
-    """Load triggers from triggers.ini into a TriggerDef instance.
-
-    Parameters
-    ----------
-    fname : str | Path
-        Path to the configuration file. Default to ``'eeg_flow/config/triggers.ini'``.
-
-    Returns
-    -------
-    triggers : dict
-        Trigger definitiopn containing: standard, target, novel.
-    """
-    fname = ensure_path(fname, must_exist=True)
-    config = ConfigParser(inline_comment_prefixes=("#", ";"))
-    config.optionxform = str
-    config.read(str(fname))
-
-    triggers = dict()
-    for name, value in config.items("events"):
-        triggers[name] = int(value)
-
-    # verification
-    keys = (
-        "standard",
-        "target",
-        "novel",
-    )
-    for key in keys:
-        if key not in triggers:
-            raise ValueError(f"Key '{key}' is missing from trigger definition file.")
-
-    return triggers
-
-
 def create_config(
     xdf_folder: Union[str, Path],
     derivatives_folder: Union[str, Path],
